+++ patches/src/main/kotlin/hoodles/morphe/patches/duolingo/misc/integrity/Fingerprints.kt
@@ -4,16 +4,14 @@ import app.morphe.patcher.Fingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
 
 object LoginStateFingerprint : Fingerprint(
+    definingClass = "/LoginRequest\$EmailLoginRequest;",
     accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.CONSTRUCTOR),
-    returnType = "V",
-    strings = listOf(
-        "password",
-        "signal"
-    )
+    returnType = "V"
 )
 
 object BasicLoginFingerprint : Fingerprint(
+    definingClass = "/LoginFragmentViewModel;",
     accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.FINAL),
     parameters = listOf("Ljava/lang/String;", "Ljava/lang/String;"),
-    strings = listOf("password", "distinctId")
+    returnType = "/LoginRequest\$EmailLoginRequest;"
 )